# package/backend/dist/swagger.json
{
	"components": {
...
		"schemas": {
			"UserCreatedDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"nickname",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseDTO": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RoleDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"isValid": {
						"type": "boolean"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/RoleDTO"
					}
				},
				"required": [
					"id",
					"nickname",
					"email",
					"image",
					"isValid",
					"roleId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateUserDTO": {
				"properties": {
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"isValid": {
						"type": "boolean"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"nickname",
					"email",
					"password",
					"isValid",
					"roleId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UsersPaginatedDTO": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserDTO"
						},
						"type": "array"
					},
					"lastPage": {
						"type": "number",
						"format": "double"
					},
					"totalRecords": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"hasMorePages": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"lastPage",
					"totalRecords",
					"currentPage",
					"hasMorePages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"isValid": {
						"type": "boolean"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "NodeJS API for LosDelFondo WebApp",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreatedDTO"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsersPaginatedDTO"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/users/{id}": {
			"get": {
				"operationId": "GetUserByID",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}